move more cryptography to cryptography.ts
